requirements.txt
openai
gitpython
radon
bandit
